# Fix apply_bit_mask cuda implementation